Functional dungeon pathfinding implemented

+++ Pathfinding/Navigation/Include/recastnavigation/Detour/Source/DetourAlloc.cpp
@@ -19,12 +19,12 @@
 #include <stdlib.h>
 #include "DetourAlloc.h"
 
-static void *dtAllocDefault(int size, dtAllocHint)
+static void* dtAllocDefault(size_t size, dtAllocHint)
 {
 	return malloc(size);
 }
 
-static void dtFreeDefault(void *ptr)
+static void dtFreeDefault(void* ptr)
 {
 	free(ptr);
 }
@@ -32,13 +32,13 @@ static void dtFreeDefault(void *ptr)
 static dtAllocFunc* sAllocFunc = dtAllocDefault;
 static dtFreeFunc* sFreeFunc = dtFreeDefault;
 
-void dtAllocSetCustom(dtAllocFunc *allocFunc, dtFreeFunc *freeFunc)
+void dtAllocSetCustom(dtAllocFunc* allocFunc, dtFreeFunc* freeFunc)
 {
 	sAllocFunc = allocFunc ? allocFunc : dtAllocDefault;
 	sFreeFunc = freeFunc ? freeFunc : dtFreeDefault;
 }
 
-void* dtAlloc(int size, dtAllocHint hint)
+void* dtAlloc(size_t size, dtAllocHint hint)
 {
 	return sAllocFunc(size, hint);
 }
