+++ Pathfinding/Navigation/MoveMap.h
@@ -35,7 +35,7 @@
 #include "Utilities/UnorderedMapSet.h"
 
 //  memory management
-inline void* dtCustomAlloc(int size, dtAllocHint /*hint*/)
+inline void* dtCustomAlloc(size_t size, dtAllocHint /*hint*/)
 {
 	return (void*)new unsigned char[size];
 }
